remove flags from path using 'remove' command with: -s -a and/or -o
overview: if removed -s, inherit from default. that's the 'removal' process

if all 3 flags are present (-s -a and -o), remove all flags from path and just keep the path

if at least one flag from -s -a or -o is specified, it will 'remove' what was specified (replace with default flag field value) and keep the original value of those not specified.
there will be a check to see if the replaced flags are equal to the default flag fields. if all are equal, just remove the flags like above. if not, keep the changes