feat(command): new run command struct
TODO: change how changing the background works. the current
implementation uses one big for loop with a nested for loop inside a
nested for loop. it is convoluted.

idea: use a struct to store the state and use events (channels) mapped
to methods to change image (or do some other action)