feat(tbg): move functionality that interacts with Windows Terminal settings.json to another file
also split up the one big function so it is easier to both understand
and maintain

NOTE: tbg refers to the process that is changing the background image of
Windows Terminal. All functionality related to that will fall under
here. On that note,...

TODO: rename change_wt_bg.go --> tbg.go