feat(tbg): implement new random behavior
OLD BEHAVIOR: on initialize, randomizes the paths and then randomizes
the images of current path. you'd need to exhaust the images in the
current path still to go to the next path.

now, the next image will be a random image from a random path. this
means that the only command available for the user is next image.